+++ README.md
@@ -1,6 +1,6 @@
 # dnd_mud
 Text game base on table-game mechanic Dungeons and Dragons.
 
-Mechanic based only on Player HandsBook D&D 5 edition 2024 year
+Mechanic based only on Player HandsBook D&D 5.5 edition 2024 year
 
 `python version = 3.10`
