--- player.py
@@ -1,7 +1,0 @@
-from dataclasses import dataclass
-from creature import Creature
-
-
-@dataclass
-class Player(Creature):
-    race: str
